chore: define and instantiate Utility class, rework type imports
